pyproject.toml
[tool.poetry]
name = "dio-bank"
version = "0.1.0"
description = ""
authors = ["jamilcfilho <jamilabcf@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "*"
flask-sqlalchemy = "*"
flask-migrate = "*"
flask-jwt-extended = "^4.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
